move code to km-document-attachment
