+++ app/api/km-document.js
@@ -18,18 +18,4 @@ export default class KmDocumentApi extends ApiBase
         return this.http.get(`api/v2013/documents/${identifier}`).then(res => res.data).catch(tryCastToApiError);
     }
 
-    async createTempAttachmentSlot(params){
-        return this.http.post(`api/v2015/temporary-files`,params).then(res => res.data).catch(tryCastToApiError);
-    }
-
-    async uploadToTempSlot(slotUrl, file, contentType)  { 
-        return this.http.put(slotUrl, file, {headers: { 'Content-Type': contentType }})
-                        .then(res => res.data).catch(tryCastToApiError);
-    }
-
-    async persistTemporaryAttachment(identifier, guid, fileName)  {
-        return this.http.post(`/api/v2013/documents/${encodeURIComponent(identifier)}/attachments/persist-temporary/${encodeURIComponent(guid)}`, {fileName})
-            .then(res => res.data).catch(tryCastToApiError);
-    }
-
 }
